Add memory size multiplier
In HIP, we are not doing the full set of reductions within the main
nbnxm kernel, but instead fully reduce in a separate kernel afterwards.

This means we need to have more memory space available for the reduction
buffers, which size we control with the new variable here.

This is one for all other implementations, but can be changed in the
future if we identify it as beneficial.

Part of #4957